requirements.txt
# Vector DB & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0

# NLP & Analysis
umap-learn>=0.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Text Processing
pypdf>=3.15.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Visualization (for figure generation)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
